Refactor: Move component to Table directory #6
