+++ src/modules/clients/components/Table/ClientsHeader.tsx
@@ -1,9 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
-import { Button, Flex, Select, Typography } from 'antd';
+import { Button, Flex, Select, Spin, Typography } from 'antd';
 import { debounce } from 'lodash';
 import { useCallback, useMemo, type Dispatch, type SetStateAction } from 'react';
 import { Fragment } from 'react/jsx-runtime';
-import { searchClientsQueryOptions } from '../clients.queries';
+import { searchClientsQueryOptions } from '../../clients.queries';
+import { LoadingOutlined } from '@ant-design/icons';
 
 type Props = {
     searchingQuery: string,
@@ -20,7 +21,7 @@ export function ClientsHeader({
     setCurrentPage,
     setIsCreateModalOpen,
 }: Props ) {
-    const { data: searchData } = useQuery(searchClientsQueryOptions(
+    const { data: searchData, isPending } = useQuery(searchClientsQueryOptions(
         { search: searchingQuery }
     ));
     
@@ -58,7 +59,7 @@ export function ClientsHeader({
                     size='large'
                     placeholder='Digite o nome do cliente'
                     style={{ width: '400px' }}
-                    notFoundContent="Nenhum cliente encontrado"
+                    notFoundContent={isPending ? <Spin indicator={<LoadingOutlined />} /> : null}
                     onSearch={debouncedOnSearch}
                     onChange={onChange}
                     options={options}
